apps/local-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
python-dotenv==1.0.1
paddleocr==3.4.0
rapidocr_onnxruntime==1.4.4
Pillow==11.1.0
numpy>=1.24
